* Added the swap functions to the list of globals in toc.cyc -- I believe the changes Mike and I were making influenced these and this is what caused the unexplained changes in the previous couple of check-ins.
* Fix to tovc so that it emits prototypes for functions only when needed
(i.e., when it has to lift out a statement expression as a separate
function.)  That's why all of the genfiles changed so much this time.

* Fixed throwing of Match_Exception's -- we weren't passing the right
thing to throw (should've passed an address) and as a result, would
get a core-dump when we tried to print out the uncaught exception's
name.  Also, I centralized this as a function call in the runtime, as
we've done for the other compiler-generated throws, which should make
it easier to catch.

* Added minimal support for asm (and __asm__).  Basically, the lexer
just snarfs in the whole thing and treats it as a string.  Useful for
"C include" of header files.  The longer term goal is here is to allow
us to parse any GCC header so that "C include" becomes useful for
including things like <kernel.h> or <windows.h> which use all of the
funky GCC'isms.

* Fixed bugs with -idirafter and related command-line arguments.  This
required me to muck with how arguments to cpp are quoted.  In particular,
the old code used to quote the cpp argument when it had a space.  I
no longer do this quoting (unless there are other special characters.)
I'm not sure it was a good change.  But without the change, the flags were
useless as implemented.