I've checked out the compiler and tests and tests/shootout, but you should try out other code with these changes.
* Fixed major bugs in subtyping and coercions.  We were essentially
letting you do deep subtyping regardless as to whether things were
read-only or not.  In turn, this required some hacking on the dictionary
file which was using the deep subtyping (had to insert a few more
consts here and there) and a few more of the compiler source files
but not much (phew!)

* Fixed bug in check_writeable -- we were allowing a struct (or tuple
or datatype field or array) to be overwritten even when it had a
component that was const.  (C does not allow this and we can't
either.)

* We now allow type definitions for structs, unions, and datatypes
anywhere that a type can go.  For instance, you can now write:

  void foo() {
    struct Foo {int x; int y;} f = {0,1};
    ...
  }

This brings us a bit closer to C and subsumes the parser hacks
we had to support:

  typedef <defn> foo_t;

where we could only support a single definition within a typedef.
However, we do not yet respect the scoping rules of C.  In particular,
in the definition above, struct Foo is supposed to in scope only
throughout the definition of the funcion foo().  However, we end up
treating it as defined within the current namespace.  The way this
works is to have an extra type_t constructor TypeDefnType that
contains definitions and a type_t* within it.  Initially, the type_t*
is NULL.  During elaboration, we enter the definition into the
environment (imperatively -- hence the scope issues) and then insert a
a pointer to an instance of the definition as the type_t*.  The
compress(-) routine looks for the type instead of the definition.

We could use this to now get rid of anonymous aggregates (and enums)
since they would be treated as "definitions" (if we gensym'd a name
for them.)  However, since we've already got the code for these
features, I'm leaving themin.

* Allow an empty name on a union field without having to have a
bitwidth (a GCC extensions).

* Allow "public" scoped types to merge with extern "C" types without
error.