Updated the region profiler to track the unique region.  I had to make the following hack to get this to work:
If malloc() or new() is called without a region argument, and
it is discovered the unique region should be used, we insert
the constant expression referring to the unique region.  This is so
that region_malloc/region_new will be called, rather than cycalloc
(which assumes the heap).  In this case, we store relevant
information based on the region handle.

To do this, I can to create a constant expression for the unique
region that must correspond with the declaration in core.h.  I
also updated the tree-shaking stuff to not prune the unique_region
extern declaration in this case.

The region profiler does not track individual free() events for
the unique region; I may add this in later if it seems useful.