* Fixed bug Bart found in error messages generated for inexhaustive pattern matches involving datatypes.  Still need to do a better job of explaining overlapping patterns.
* We now allow declarations such as "register i;" which do not have
a type specifier.  However, we issue a warning.

* Fixed a bug James found where we weren't getting qualifiers right
in a declaration with multiple declarators as in: const char *x, *y;
(James fixed this too, I think.)  Still need to get const right for
return types on functions.

* Fixed a bug James (and Dan) found where we couldn't use a typedef
in an argument position when that typedef was an array type.  For
instance:

typedef float vector[3];
void foo(vector v);

This is now elaborated to:

typedef float vector[3];
void foo(float@{3}`r v);

where `r is a fresh region variable.  Note that in this situation,
there's no way to explicitly put a region in for the "array" type.

* Added some context to the environment in toc.cyc which lets
us know when we're in a left-hand-side expression.  I'm hoping
to use this to get rid of the bug Mike found in assignments of
the form x.y.z = e where both x and y are tagged unions.