The allocation profiler now correctly tracks unique pointer objects that are garbage collected.  (refcount objects are still to be fixed.)
This is done by registering a gc finalizer with each
unique pointer object; the finalizer prints its
address if the gc frees the object.  Each allocation
now prints the address as well (to amon.out).  Thus
we can track the exact lifetime of any object.
Currently this is only used for unique pointer objects
but it could be done for all objects, giving a
producer graph (and that's next...)