My last checkin forced me to use the %pure directive in the parser since we are region-allocating the parser's temporary data and thus communication with the lexer for semantic values needs to be done in some way other than a global.  Unfortunately, %pure also has the effect of making yylloc a local variable to the parser and since its contents is bits-only, it was never being written with the right values (rather, the lexer was writing the global variable.) Thus, locations were screwed up.
The right fix is to change the lexer so that it takes &yylloc as
a parameter, but I haven't done this yet.  Instead, I've temporarily
commented out the declaration of the local yylloc within yyparse
in cycbison.simple when %pure is specified.  I need to fix the
location stuff to get this right so that our parser really is
"pure".