When you compile with the -compile-for-boot flag, then the file names that get inserted into the generated C code for the line directives have their paths stripped off.  This will (hopefully) get rid of any absolute path names in bin/genfiles which would've caused us to be doing updates all the time.
Unfortunately, this makes the line directives a bit less
useful for the installed libraries.  For instance, gdb
won't know where to find things.  Oh well.

Also, changed absyndump to not put in newlines (it was
doing this to make it a bit easier to view the output)
so that the linenumber info matches up with the generated
directives.