Fixed the first of Mike's bugs where he had:
  const char foo[3] = "foo";

Note that the length should be 4 (due to trailing zero which
always comes with string constants).  That was causing us to
insert a cast.  We no longer put the cast in, but you should
really change the length to 4 :-)

Other changes are in preparation for allowing a struct to
be terminated by an array of indeterminite length.  This is
not yet working.