Added support for declaring noconsume(x) arguments for Cyclone functions. This attribute is used to indicate that the function will not free (consume) the named argument.  This is checked very conservatively at the moment: the location pointed to by the parameter upon entry must be the same upon exit.  This could be relaxed later to account for assigning the location to some non-escaping local variable and/or dropping the pointer.
I implemented this by adding another absRval type to the flow analysis for a
"named location:"

    NamedLocation(string_t<`r> name, absRval_t<`r> actvalue);

The idea is that when a function is called, it points to some named location
(call it \rho).  As the function executes, the initialization status or
abstract value at this location might change, but the name won't.

The manipulation of NamedLocations by the flow analysis is basically
intuitive.  For example, we only have a named location as the result of a
join if the two values have the same name; else we throw away the name.
Also, by convention, we consume the actvalue of the NamedLocation, rather
than the NamedLocation itself (i.e. we can have
NamedLocation(n,Consumed(...)) but not Consumed(NamedLocation(n,...)...));
this is like Aggregates.

At the moment, initializes(), noliveunique(), and noconsume() attributes are
mutually-exclusive, though some combinations would be sensible.  Also,
initializing a let-bound variable with a named location fails to propagate
the name (or any other information about the absrval, for that matter; this
is because lets generalize to patterns and we don't special case when the
pattern is just a variable).

Also fixed code for absynpp.cyc to be in line with absyndump.cyc
for printing temporary tvars.