* Changed named locations to use a vardecl for the name.  This allows me to uniquely index the rvalue with something that's stable across iterations of the analysis.  Without this, we were losing the identity when we had an outer iteration inside of which was something that needed to name an rvalue (e.g., in the alias construct -- see below.) For instance, the do loop that contains an alias declaration in lib/queue.cyc did not type-check without this change.
* I've now "fixed" the alias construct (I hope -- needs to be modelled
and proven correct.)  But, the changes to support this were
suprisingly easy.  In particular for:

 { alias <`r> T x = e; s; }

the flow analysis ensures that the location to which e evaluates
cannot be used within S.  It also ensures that this location does not
get overwritten if it's accessible from a local variable.  This rules
out the bad case where you write:

  { alias <`r> T x = y->tl;
    y->tl = unew List(3,NULL);
    ufree(y->tl);
  }
  y = y->tl;  // core dump!

I use named values in the analysis to make sure the value in the
aliased place is the same upon exit from the block.  We could do
something weaker (eg., if it's not the same, then do not update the
context with the original initialization info for the place) but this
was the easiest thing to do for now.

Second, when type-checking, we ensure that the the type of e is a
subtype of T, under the assumption that the unique region
outlives `r.  This lets us do something like the following:

  struct CList<`a,`r> { `a hd; struct CList<`a,`r> *`r const tl; };
  typedef struct CList<`a,`r> *`r clist_t<`a,`r>;

  int clength(clist_t<`a,`r>);

  void foo(list_t<`a,`U> x) {
    int i;
    { alias <`r> clist_t<`a,`r> y = x;
      i = clength(y);
    }
    freelist(x);
  }

Note that the subtype relation holds only because the tail fields are
read-only. If instead, you tried to declare y to be a normal list_t,
it wouldn't type-check.  This prevents the bad case where you do
something like this:

  { alias <`r> list_t<`a,`r> y = x;
    y->tl = y;  // x->tl = x -- circular list now
  };
  let z = x->tl;  // z = x, x = x -- aliases!
  ufree(z);
  x->hd;   // core dump!

Even cooler is that Mike's got code that inserts the alias
statement when a function call fails to type-check and re-tries.
So, you can actually wrie foo above as follows:

  void foo(list_t<`a,`U> x) {
    int i = clength(x);
    freelist(x);
  }

Note, however, that you do get a warning that the call to
clength(x) is coercing x from a list_t to a clist_t.  I had
to do a bit of hacking on the alias-insertion code to get
this particular example to work out, but it does, and I'm
very pleased.

Now, if only we had const polymorphism...