Added support for dynamically sized, flat arrays in structs (as the last field).  Now you can do
ptr_t<int> vectorn(tag_t n, int x) {
  return new Ptr{.size = n, .elts = {for i < n : x}};
}

This is implemented as follows.  For typechecking, we allow
comprehensions for the last field of an abstract struct that is
itself being dynamically allocated, and munge the type to
keep things happy.  However, this last field cannot itself
be an abstract struct with an array in it (though we might like
that), because of code generation issues.  I added a field to
tcenv to pass information downward so that the typechecker
will permit a dynamic sizing on the comprehension in the
acceptable context.

For code generation, we simply change the malloc() expression
for the struct to be (in the above example)

malloc(offsetof(ptr_t<int>,elts)+sizeof(int)*n)

and then initialize everything as before.  The limitation of
offsetof to a single field prevents taking the offset of the array
part of a nested aggregate; hence the limitation mentioned above.

Implementing this revealed a bug in tovc due to mutating variables to
be boxed when generating a function from a statement, which I fixed
by doing a little more copying.  I also fixed a bug in the parser
that resulted in not properly setting the scope of some type declarations.