* Fixed bug Mike found in top-level (polymorphic) struct initializers. We weren't inserting casts where needed.  Now, what I do is compare the expression's translated type against the field's translated type. If they don't unify, and if they aren't arithmetic types, I cast the expression to the translated field's type.
* Fixed bug where we weren't deleting foo.cyp when there was an
error in pre-processing (thanks to Mujtaba for reporting.)

* Changed xarray's to use a unique pointer to the underlying array.
That allows us to free the old array when we have to grow.
The reuse() operation also frees the underlying array.  Note
that the struct that holds the array pointer and the size still
go into a region.  So, the client is responsible for calling
reuse(), else the underlying array will not be reclaimed when
the region goes away (rather, it will be reclaimed by the collector
if it's linked in.)

* Changed pp.cyc and toc.cyc to call Xarray::reuse when they're
done with their extensible arrays.