Warning:  this change will force Cygwin users to upgrade to the latest version of Cygwin 1.3.22-1.
Cygwin now seems to get the definition of __uint16_t and such
right (at least consistent with Linux) so I eliminated some
of the #ifdef's in libc.cys that were attempting to work around
this.  We were trying to define these, but they're now built in
so we'd get an error when compiling the generated cstubs.c
Because older versions of Cygwin do not define these things
properly, and they're not built in, it's likely that cstubs.c
will fail to compile.  Rather than try to maintain two machines
with two different versions of Cygwin so we can test, I suggest
everyone just upgrades.

Also, added flatlist.cyc to the tests directory (which I thought
I had done earlier) and folded it back into the Makefile.