Reworked how we do bounds checking.  First, I cleaned up the code that tracks the relations in the flow analysis quite a bit, and generalized it so that we can track ==, !=, <, <=, >, and >= for constants, non-escaping variables, numelts of a non-escaping variable, and an integer-kinded type variable.  It's relatively easy to extend these relations with new kinds of operands (e.g., if we wanted to track i <= x+1.)  Dan, you'll be happy to note that I've finally gotten the functionality you used to have folded back in.  In particular, if you compare (say) x <= e, and e happens to have type tag_t<`i>, then the relations track the fact that x <= `i.  (See anal_primop_test in new_control_flow.cyc, along with a bit of cleverness in CfFlowInfo::exp2relnop().)
Second, I added a simple implemenation of Pratt's algorithm for
determining whether a set of constraints of the form x - y <= c is
satisfiable (see pratt.cyc).  It's possible to encode all of the
relations we track using these constraints, and for those constraints,
Pratt's algorithm is complete.  My implementation is pretty naive (as
in cubic running time) but it doesn't slow down the compiler (there
are very few constraints, they're sparse, and we only call this on
demand for a subscript operation instead of on each flow step).
Regardless, it's abstracted out so that it's easy enough to change the
data structures and reduce it to a quadratic algorithm if this ever
becomes a problem.

So, this makes the bounds checking a little less brittle than it was
before.  In particular, Bart's code:

void foo(tag_t n, int arr[n]) {
  if (n >= 3) {
    arr[0];
    arr[1];
    arr[2];
  }
}

will now type-check.  The one thing we still don't cover well is
counting down from numelts(x)-1 to 0.  The key problem is integer
underflow/overflow, which is the primary reason why I didn't implement
some more sophisticated integer linear program solver.

I should say that the treatment of joins for the relations is still
ad-hoc and largely syntactic.  We could use Pratt's algorithm to find
the maximal distance between any two variables as part of the join,
but this ordering results in an infinite chain, so we wouldn't reach a
fixed-point.  An alternative is to use Pratt and keep a finite
distance between points if they are the same on both paths, but drop
the constraint otherwise.  This is essentially what we're doing now
(but without invoking Pratt.)

I now feel a little less embarassed about how this works :-)