Factored the relation stuff out into its own files since I now need it elsewhere (see below).
I desperately need an infrastructure for regression testing
things that should *not* type-check, if someone gets some
spare cycles.

Added support for type-safe untagged unions which I will call,
for lack of a better term, "require" unions.  A test file
is in tests/require.cyc.  One can write:

union Foo<`i> {
  int *x  @requires valueof(`i) >= 3;
  int  y; @requires valueof(`i) < 3;
};

If f is a value of type union Foo<`i>, then you have to be able to
prove that the requires clauses are satisfied to read or write a given
member.  For instance, one can write:

int g(tag_t<`i> tag, union Foo<`i> f) {
  if (tag >= 3)
    return *f.x;
  else
    return f.y;
}

A datatype can be encoded by pairing a tag and union in an existential:

struct FooWrapper {
  <`i>
  tag_t<`i> tag;
  union Foo<`i> value;
};

The requires clauses for a given union must be static expressions
(i.e., only include constants, operations like +,-,*, etc. and
valueof(`i) where `i is an Integer-kinded type variable.)  For
now, they are further restricted to simple comparisons of
the form:  R1 && R2 && ... && Rn, where each Ri is a relation
of the form Rop1 Reln Rop2, and Reln is drawn from {==,!=,<,<=,>,>=}
and Rop is in {<constant>,valueof(`i)}.

Furthermore, the requires clauses for the members of the union
must be non-overlapping.  So, for instance, it is not possible
to write:

union Foo<`i> {
  int *x  @requires valueof(`i) > 3;
  int  y; @requires valueof(`i) < 3;
};

because when `i == 3, we could write an int and read an int*.  We use
the constraint solver to check that the relations don't overlap.

The flow analysis context determines whether or not a @requires clause
is satisifed, just as array bounds checks are determined.

In the future, I hope to add more support for more operations (e.g.,
bitfield twiddling?) as I think we'll need this for real code.  Also,
I'll probably make it possible to add @requires clauses after the
fact.  This way, you could (in principle) suck in a C interface (which
doesn't have the @requires clauses) and then specify them later.  I
may also make it possible to read out bits-only types if the @requires
clause is not satisfied, but issue a warning.

Finally, I also hope to add @requires clauses to functions and
existentials which will extend the relational stuff to the
inter-procedural world.  We could also think about adding support for
@ensures clauses for functions.  The cool thing is that most of
the necessary hard stuff is in place now, and it's factored to
a degree that it should be possible to add more smarts in an
incremental fashion.