Added code to interpret patterns in the flow analysis.  The upshot is that we can now match against expressions that aren't fully initialized.  This is leading towards being able to properly track uniqueness through pattern matching.
The basic change is that rather than associate a list of pattern
variables in Let_d and Switch_clause elements, we track variables
and corresponding expressions.  These expressions are generated
from the pattern and the expression being matched.  For example, if
I had

struct Foo { int x; };
struct Bar { struct Foo f; int y; };
void f(struct Bar @b) {
  let &Bar{.f=fvar,.y=yvar} = b;
}

Then the pattern variable list would be

[ $("fvar",(*b).f); $("yvar",(*b).y) ]

(where "fvar" and "yvar" are actually vardecls).  Since we can't
generate similar expressions from datatype pattern matches, these
are still uninterpreted.  To ensure they are initialized, we
additionally note a list of expressions that lead to datatypes in
patterns, and the flow analysis "uses" these.  This is described
more precisely in tcpat.h.

In order to generate the pattern variable expressions, I require that
either the pattern is a variable, or that the expression being
matched against is a variable.  If neither condition is true, I
insert a declaration that assigns the matched expression to a
variable first.  Unfortunately, for some reason this forced me to
do a deep copy when generating the init expressions (else things
break, not sure why), which apparently mucks the flow analysis a bit
so that some null checks that were eliminated before are not now.
Perhaps someone more knowledgable here could take a look?

There is code now to check uniqueness properly in patterns, but
the flow analysis is not yet fully functional in this regard.  I'm
working on this part next.

Also fixed a small bug in tovc in which AnonEnums were being
neglected.