Updated typechecker to validate pattern matching on unique pointers. Essentially we do two things: 1) For each access expression generated per pattern variable, we check that if it's a unique pointer, that the access path is unique. 2) For each reference pattern, we check that the corresponding access expression is not taking the address of a unique path. At the moment, we can finesse worrying about bad paths/patterns within datatype constructor patterns, because these constructors are not allowed to contain unique pointers (but this should ultimately change).
Also fixed the flow analysis to "expand" unique locations early so as
to permit more code to go through.  Say we have something like

struct foo { int @`U x; int y; };
void f(struct foo @`U f) {
  Core::ufree(f->x);
}

Before, this would fail because in the flow analysis, f is initially
mapped to Unk(all-init).  During the unordered expression evaluation
induced by the function call (considering expressions ufree and
f->x), f->x would be "expanded out" to create an abstract location
for its x field, so its "consumedness" could be tracked.  I realize
now that this can be thought of as akin to the isomorhpic expansion of
a recursive type.  However, when this expansion takes place in
an unordered context, we end up joining the original flow with
the expanded one after the first iteration, and lose information.
The result is that f->x is considered uninitialized.

To fix this, I just expand the flow early, just before the unordered
iteration takes place.  This is sound for the same reason that
expansion is sound in the first place: the expanded and non-expanded
flows are isomorphic because the path being expanded must be
completely unique.

I added a bunch of test cases to patterns.cyc to check all this stuff.