Fixed a bug that prevented us from getting NULL checks right on copies of unique pointer values that are cast from @nullable to @notnull.
Fixed a bug that prevented us from using alias on a parameter that is
marked as noconsume.  This also fixed the bug that allowed:

int f(int *`U x) {
 { alias<`r> int*`U y = x;
   ufree(y);
 }
 return *x;
}

We now check that a break doesn't cause a function to fall off in the
flow analysis instead of the type-checker.  This makes it a bit more
accurate (and thus fixes a "bug" found by Rajit.)

Fixed a bug Bart found in the location for a declaration with no body.
We were using a "0" for the location of the whole declaration instead
of the actual location.