The major change is that I've added the Doug Lea allocator to the runtime system, so that when you compile with -nogc on a Linux system, you get that instead.  The benefit is that some of the profiling commands that it provides are now accessible to aprof.
If you want to try to compile things on a non-Linux platform,
feel free to update the ifdef in nogc.c to include that platform.

Fixed the textual summary of aprof to include the refcnt region,
and fixed runtime accounting of refcounted objects (was getting
the wrong size).  Now allow noconsume() attribute on non-unique
pointers, for polymorphism purposes.