Fixed bug that Trevor found with alias.  In tcstmt, check_alias_coercion was inserting an unchecked coercion on the "e" in alias <`r> T x = e changing e's type to be x's type.  With e now no longer being unique, it was not being consumed by the flow analysis, so you could free it inside the alias block.
I "fixed" this (and the ripple-effect problems in new_control_flow.cyc)
by removing the unchecked coercion, thus leaving e's type intact.
However, I think this is not correct in general, as T could be
something like int *`r while e's type could be int ?`U.  This implies
a computational change that the coercion would signal.  Instead, we
need to ensure that these types are either isomorphic, or that we
can perform the coercion while leaving the `U's in place.  Can someone
else (Greg?) take a look at this?

Interestingly, this fix revealed some bugs in the streambuff.cyc test
file, which came from MediaNet.