Removed the Alias_d decl construct in favor of a special alias pattern AliasVar_p(tvar_t,vardecl_t).  Therefore, what was once
alias <`r>T x = e;

is now simply

let alias <`r>T x = e;

Alias variables can more generally appear in nested pattern locations, in
switch cases, etc., but not within a datatype pattern, because we cannot
create an non-pattern expression (like x->y.z) to designate the variable's
initializer.  Alias is no longer a keyword.

In addition to the flexibility and code reduction benefits of doing this, it
nicely consolidates the different mechanisms for unconsuming unique pointers
that were used in the flow analysis.  In particular, the flow analysis now
attempts to unconsume any unique pointers that were matched against in a
pattern at the conclusion of the pattern block.  This is allowed to occur
when

1) The expression that initialized the pattern variable has not changed
(i.e. been overwritten or consumed).

2) If the pattern variable is a normal variable, then it must not have
changed either.  If the pattern variable was an alias variable, then its
final value is immaterial.

We use named locations to implement these two conditions.  We restore the
original to the join of the pattern variable and the initialization
expression when the conditions are met.

For type checking, I had to augment pattern matching results to include
regions `r introduced by alias <`r>, and then modified pattern_synth to
include those regions during checking.  This is different than opening an
existential region because the region is assuredly "live," but is otherwise
similar.

I checked in a test file unique_patterns.cyc that tests this stuff out
thoroughly, for patterns and alias patterns.  I'll update the fail*.cyc
files for more negative tests shortly.

There are still some situations in which alias variables are not allowed
when they should be; I need to do some type munging to get this to work.
Also, I still need to do unconsuming properly when entering exception
handlers.  Parsing isn't too pretty right now: basically the parse line is

'alias' '<' type_var '>' type_name IDENTIFIER

The problem is that "type_name IDENTIFIER" doesn't work too well for arrays,
function pointers, etc.  Putting a declaration here would be better, except
that we don't want an initialization expression or a semi, and in any case
this results in lots of shift/reduce problems.  If someone has ideas here,
I'd be happy to change this.