New, simpler form of dynamic regions that in essence gives us unique regions and reference-counted regions (as opposed to unique and reference-counted pointers.)  This interface has lots of advantages including the fact that handles no longer have to leak, and we don't have to add any new constructs to the language, rather only a few run-time routines with some very advanced types (which we can write in the language without any changes.)  Details are below.
Fixed a bug in Toc::set_lhs_exp -- used `r where we should've used `r2.

Modified subset_effect so that it does a little better job inferring
instantiations for effect variables.  I'm using them for the new
treatment of dynamic regions, and wanted to avoid writing explicit
effect instantiations.  This change triggered an error in
Toc::set_lhs_exp because there was a region ordering constraint on a
free region variable (`r2) which got instantiated differently as a
result of the change.

Details on new dynamic regions:

I've left the old dynamic interface stuff in there for now, but warn that
it is deprecated in favor of the new interface.  Please switch to the
new interface if you're using it in any applications.  I've already
modified the compiler and checked that all the tests still run.

The new interface is quite simple and lives in core.h.  Basically, we
represent a dynamic region as a unique (or reference-counted) pointer
to an abstract struct DynamicRegion (which just contains the region handle).
The unique (or ref-counted) pointer is called the key.

The operation new_ukey() creates a fresh region and returns a unique
key for the region.  (new_rckey() creates a fresh region and returns a
ref-counted key for the region.)  The operations free_ukey() and
free_rckey are used to destroy unique and ref-counted keys
respectively.  Because ref-counted keys are just `RC pointers, you can
use alias_refptr to make a copy of a ref-counted key.  Obviously, you
can't make a copy of a unique key.  The free_ukey() operation reclaims
the key's region, as well as the storage for the key.  The
free_rckey() operation decrements the reference count, and if it's
zero, reclaims the key's region as well as the storage for the key.
Note that you can pass a ref-counted key to drop_refptr and you can
pass a unique key to ufree, but doing so doesn't actually deallocate
the region, but rather only the key.

The operation open_dynamic_region(k, arg, f) extracts the handle h
from key k, and returns the result of f(h,arg).  The operation does not
consume the key.  Note that for unique keys, there's no way for
f to get its hand on the key for the region (e.g., to free it).
This is because we already passed in one copy of the key to
open_dynamic_region() and you can't make copies.  For ref-counted
keys, you can make copies, but that ups the reference count.
Thus, we are guaranteed that the region will be around throughout
the call to f.

The interface for open_region is as follows:

extern `result open_region(struct DynamicRegion<`d> @`r::TR key,
                           `arg arg,
                           `result body(region_t<`d> h,
                                        `arg arg;{`d,`r}+`eff);
                           {`r}+`eff) __attribute__((noconsume(1)));

There are a few things to note.  First, you can pass in a `U or `RC
key to the open construct, so we only need one open for both kinds of
dynamic regions.  Second, note that the caller of open_region() does
not need the dynamic region (`d) to be in its capability, even if
body() does.  This is because the key is evidence enough that the
region will live throughout the call to body().  To give the caller
enough flexibility, I had to use an effect variable to specify the
capability.  Here, `eff represents the set of regions that the caller
of open_region knows are live except for `d and `r.  Note that `r (the
key's region) must be live, but that must be either `U or `RC, so it's
never a problem.  (There's no reason we couldn't support keys in
other regions, it's just that there would be no way to destroy
them, except say through GC.)

You can "share" a dynamic region key by placing it in some shared data
structure.  Of course, you'll then have to swap with NULL to get it in
and out of the shared data structure.  The `U keys work fine for this
except when you need to open the same region multiple times.  The
problem, of course, is that if you swap in NULL, then whoever tries to
swap it out will fail.  In other words, you can't really do recursive
opens with `U keys.  However, you can do this with `RC keys!  Swap out
the key, make a copy of it, swap it back in, and use the copy for the
open (making sure to destroy the copy after the open).  Thus, an `RC
key provides just as much functionality as our original dynamic
regions.

If you want to see examples of these in use, then look at lex.cyl and
toc.cyc.  The former uses `U keys and the latter uses an `RC key and a
clever higher-order, rank-2 polymorphic function, use_toc_state, to
abstract out all the swapping that has to go on.  It's waaaaaay
cool and effectively implements all of the functionality that
we have in the run-time for our old dynamic open mechanism, except
that it's implemented within the language (modulo open_region()
which is literally one line of C code.)

All of the other operations are implemented in runtime_cyc.c and are
just a few lines each too, though I have not yet integrated support
for region profiling with this form of dynamic regions.  I will
probably add a scoped open construct that uses the new interface to
make it easier to open keys (saves you from writing a separate
function -- oh how lambda would be nice.)

One disadvantage of this interface relative to the old one is that if
you put a key in some shared storage in a region `r, then it is not
the case that when `r is deallocated that the key will be destroyed
automagically.  This did happen with the old interface.  Now it's up
to you to do the right thing or let the GC eventually collect it.

In the long run, the right thing to do is add a finalizer interface
for regions so that we can register a routine to deallocate a dynamic
region whenever we put it in a shared data structure.  (The same goes
for any unique pointer --- we ought to have a way to register a
finalizer...)