Added support for opening dynamic regions using unique or reference counted keys.  In particular, you can write:
{ region h = open(k); ... }

where k is either a Core::uregion_key_t<`r> or a
Core::rcregion_key_t<`r>.  This adds `r to your capability throughout
the scope of the open declaration, consumes the key k throughout the
declaration, and then unconsumes it upon exit from the
declaration---well, usually.  If you happen to jump out of the scope
of the open via an exception handler or goto or break, you lose the
key.  Tough.

We should be able to rewrite the code that uses Core::open_region
to take advantage of this now.