Updated RELEASE_NOTES, fixed documentation on "as" patterns, and fixed a bug in the flow analysis.  We were doing a pattern match like this:
  case Foo(...,.x = true,...): ....
  case Foo(...,.x = _,...): ...

We should never, never, never use true/false in pattern matches.  They
are not constants!  They are variables so the first case will match
regardless of the value of x.  (That particular case wasn't exhaustive
hence the lack of warning from the compiler.)

Instead, one should always use:

  case Foo(....,.x = v,...) && x: ...

Note that it's not even a good idea to use 1 since any non-zero value
is "true" in some situations.  It is okay to match against 0, but
probably bad form given the above.

I've modified the type-checker to warn when you use "true" or
"false" within a pattern to avoid this potential problem.