Fixed bug in pattern matching (and member selection) where we were getting out a value whose type was of abstract kind (can happen when we're taking the address of the resulting path, or when we're getting at some known component while pattern matching.)  The structarray.cyc test caught this bug (wahoo!).
Running make in tests/ resulted in no problems except that the
fail<n> output is now different and needs to be updated.  I also
checked that the tests/shootout and cyclone_internal/scheme
code builds properly (as well as the compiler and libraries
of course.)

There's still some potential hacking to do on this "structarray"
stuff to clean it up.  In particular, we could now support terminating
a struct with an array of types of abstract kind (we don't currently
support this but now could -- just make them aligned, as with other
abstract types appearing at the end of a struct.)