Added @requires clauses for functions.  For example, you can write:
  int f(int x, int y, int z) @requires(x <= y && y <= z);

The function f can then only be called when the flow analysis can show
that the parameters passed satisfy the @requires clause.  A function
defined with a requires clause gets to assume these relations on
input.  So, for instance, one can write:

`a sub(`a@{valueof(`n)} a, int i) @requires(i < numelts(a)) {
  return a[i];
}

and there will be no bounds check on the array.  The clauses are
restricted in the same way that they are for unions: They must be a
conjunction of primitive (unsigned) comparisons involving only
constant expressions, the parameter variables, or numelts(-).

Subtyping allows you to strengthen the clause.  For instance, if f
requires (x > 42), then you can assign f to a variable with function
type that has a requires clause (x > 45 && y < 37).

I also reworked a lot of the relation code so that it was a bit more
general (but not much).  This managed to get rid of a couple more
bounds checks in the compiler.

I've also added syntactic and typing support for @ensures clauses, but
have not yet folded them into the flow analysis.

I changed function types to be abstract (as they should be)
which gives us more subtyping on function pointers (don't
have to be read-only.)  This uncovered a bug in the hashtable
code (too many dereferences) which C fixes by re-promoting
to a pointer type.  Go figure.

I've given a lot of thought about incorporating signed integer
comparisons, since they tend to arise naturally.  But I'll
get to this later...