Fixed a bug in cycflex.cyl -- wasn't using the new Arg::parse correctly.
Massive rewrite of the code that does pattern match compilation for
switch and try/catch.  In a nutshell:

In tcpat we now build enough information into decision trees that we
can use them to do code generation.  We attach the decision tree to
the switch and catch to pass to the back end.  To simplify the
back-end, we also changed the treatment of catch so that the default
case (which re-throws the exception) is added by the type-checker.
See tcpat.h for the struture of the decision trees.

In toc.cyc, we use the decision tree to generate the code.  The tree
expands into a combination of switches and if-then-elses.  I use
switches wherever possible (e.g., integer and character tests,
datatype tag tests, and tagged union tests) as long as there is more
than one non-default case.  So, for instance, nothing special has
to be done to get a switch in the generated code when the original
code is a C-switch (i.e., doesn't do any fancy pattern matching.)
It's not clear that GCC emits great code for small switches, so we
may want to change the default behavior to generate if-then-else's
for small numbers of cases.  This should be easy to do.

Compared to the previous translation, we emit one fewer "goto" for
most cases (e.g., if they only show up once in the decision tree) and
do fewer tests (each sub-pattern is tested at most once along any path
in the tree---the old code would re-test overlapping sub-patterns.)
However, there is some inefficiency in how we extract values to be
tested and the variables that make up the pattern.  I should be able
to get rid of this overhead with a bit more hacking.  There is no
attempt to make the decision tree "wide" instead of "deep" (e.g., by
picking the right component to test first.)  Rather, we still end up
testing the left-most, top-most pattern first.

The performance of the compiler, which does a lot of pattern matching,
didn't improve.  This isn't too surprising since we added a lot of new
code.  The good news is that we roughly broke even with the old
compiler in spite of the fact that we're executing more code.  Then
again, compilation time is really dominated by gcc.

I have not yet treated "where"-clauses in try/catch.  The old match
compiler ignores them since it didn't need to check exhaustiveness for
catch, and since we didn't use the resulting decision tree for code
generation.

Also, I have not yet gotten rid of the original xlate_pat since we
still use it in the translation of let-declarations.  But I'm getting
ready to rip that out.  It should make toc.cyc a lot smaller and a bit
simpler.