The type-checker no longer sets the statement-annotations for the control-flow checker.  The latter does it itself.
We no longer use encloser fields on statements.  So they're gone from
the type-checking environment and the flow analysis.  Instead, the
jump-analysis computes how many patterns "with variables" a jump
exits.  The notion of "with variables" is a bit brittle:
  * For a switch-clause, we call Tcpat::has_vars (which is new)
  * For a let-declaration or an open-region, it is "always"
  with-variables.
The jump-analysis and the flow analysis have to agree on this
definition (or the unique_pat_vars part of the environment in the flow
analysis won't be right).

See NewControlFlow::unconsume_pat_vars to see where this simplified
things.

Everything bootstraps and passes the tests, particularly
streambuff_test and unique_patterns.

Now the only thing we're using statement-annotations for is the
"visited" int.