No more varmap in translation to C.  This actually simplified pattern-compilation as well as "normal" variables.
I can't figure why the Reference_p case had to set the type of the
vardecl (at quick glance tcPat seems to be doing the right thing), but
other than that everything is normal.

To summarize the change: To alpha-convert a variable you just mutate
the name field of its vardecl_t.  It does not matter if you do this
before or after translating bound occurrences.  We alpha-convert to:
 * hoist a declaration while avoiding capture
 * translate let-expressions because our scope rules differ from C's.