The main change is that I've updated buildlib to support a modicum of user customization of type definitions for buildlib.  See the updated documentation (!) for details.  I've been able to update the libc.cys file to more accurately model some C types, including struct group.  I also added support for locale.h; let me know if you have problems on non-Linux systems. There are many others to be done, but I haven't done it yet.
I also fixed a bug in which we weren't doing null-check elimination for
left-hand-sides.  Code such as

if (buf != NULL)
  buf[i] = x;

was inserting a null check for buf.

Finally, I added a function Core::mkthin for converting a fat pointer to a
pair of a thin pointer and its bound.  This is a simpler version of arrcast
that might be more useful.  I'm intending to look at changing some
libc.cys wrappers to use this.