(I got a lot of line-ending errors that I think I've cleaned up. Apologies if I didn't and there may be some "changed" files I didn't actually change.  The only real danger is I could have clobbered a recent change by someone else.  I think not, and we can roll back if I did.)
Mostly I ripped out manual memory management from type-checking
environments, making the code easier to read.  There are a bunch more
uses of TEMP_RGN and heap_region to get rid of in due course.

Also changed the ordinaries maintained during type-checking to include
only ordinaries (of type binding_t) because all we use the ordinaries
for is tree-shaking.

The change to fail.out is just shifting of evars.  Changed
tests/Makefile to use diff -w for the failtests.