Updated RELEASE_NOTES to include changes since last version. Fixed the bug that Dan mentioned with regard to function calls and expression sequencing.  Also found and fixed a bug in the translation of struct and array initializers.  In particular, given
let v = {e1,e2};

We were translating this to be

let v = ({let x1 = e1 in let x2 = e2 in {x1,x2}});

This caused gcc to barf.  So I changed the translation to lift the
brackets outside of the sequencing part, thus translating as

let x1 = e1 in let x2 = e2 in let v = {x1,x2};

Lots of scary side-effects to get this to work!  Note that this
change also cleaned up the translation to flatten some
redundant statement expressions.  Finally, I updated some of
the documentation and Makefiles.